[@dhealth/components] fix(test): update import path for refactored graphics widgets
